ValidationException. Models: remove String timestamp. Move Address to card package
