+++ fitpay/src/main/java/com/fitpay/android/api/models/card/Address.java
@@ -1,4 +1,4 @@
-package com.fitpay.android.api.models.device;
+package com.fitpay.android.api.models.card;
 
 public final class Address {
