+++ fitpay/src/main/java/com/fitpay/android/api/models/card/CreditCardInfo.java
@@ -1,7 +1,5 @@
 package com.fitpay.android.api.models.card;
 
-import com.fitpay.android.api.models.device.Address;
-
 /**
  * Created by Vlad on 11.03.2016.
  */
