+++ fitpay/src/main/java/com/fitpay/android/api/models/user/UserModel.java
@@ -18,21 +18,10 @@ abstract class UserModel extends BaseModel {
     protected UserInfo userInfo;
 
     /**
-     * ISO8601 string providing the date of the creation of original user account.   If not known use the current date
-     */
-    protected String originAccountCreatedTs;
-
-    /**
-     * ISO8601 string providing the date that the FitPay terms and conditions were accepted.
-     */
-    protected String termsAcceptedTs;
-
-    /**
      * The version of the FitPay terms and conditions that were accepted
      */
     protected String termsVersion;
 
-    protected String createdTs;
     protected Long createdTsEpoch;
     protected Long termsAcceptedTsEpoch;
     protected Long originAccountCreatedTsEpoch;
@@ -49,24 +38,12 @@ abstract class UserModel extends BaseModel {
         return termsVersion;
     }
 
-    public String getTermsAcceptedTs() {
-        return termsAcceptedTs;
-    }
-
     public long getTermsAcceptedTsEpoch() {
         return termsAcceptedTsEpoch;
     }
 
-    public String getCreatedTs() {
-        return createdTs;
-    }
-
     public long getCreatedTsEpoch() {
         return createdTsEpoch;
-    }
-
-    public String getOriginAccountCreatedTs() {
-        return originAccountCreatedTs;
     }
 
     public long getOriginAccountCreatedTsEpoch() {
